apps/users-service/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
PyJWT==2.11.0

pytest==9.0.2
pytest-django==4.11.1


mysqlclient==2.2.7